requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
streamlit==1.28.0
joblib==1.3.2
python-dotenv==1.0.0